refactor(linter): split service code into separate modules (#6437)
Refactor in preparation of multi-config/`override` support.
